updated timing, moved 'load_images'
