+++ burstphoto/utils.swift
@@ -1,4 +1,5 @@
 import Foundation
+import MetalKit
 
 
 func optionally_convert_dir_to_urls(_ urls: [URL]) -> [URL] {
@@ -16,6 +17,57 @@ func optionally_convert_dir_to_urls(_ urls: [URL]) -> [URL] {
     }
     return urls
 }
+
+
+
+func load_images(_ urls: [URL], _ progress: ProcessingProgress) throws -> ([MTLTexture], Int) {
+    
+    var textures_dict: [Int: MTLTexture] = [:]
+    let compute_group = DispatchGroup()
+    let compute_queue = DispatchQueue.global() // this is a concurrent queue to do compute
+    let access_queue = DispatchQueue(label: "") // this is a serial queue to read/save data thread-safely
+    var mosaic_pettern_width: Int?
+
+    for i in 0..<urls.count {
+        compute_queue.async(group: compute_group) {
+    
+            // asynchronously load texture
+            if let (texture, _mosaic_pettern_width) = try? image_url_to_texture(urls[i], device) {
+    
+                // sync GUI progress
+                DispatchQueue.main.async { progress.int += 1 }
+    
+                // thread-safely save the texture
+                access_queue.sync {
+                    textures_dict[i] = texture
+                    mosaic_pettern_width = _mosaic_pettern_width
+                }
+            }
+        }
+    }
+    
+    // wait until all the images are loaded
+    compute_group.wait()
+    
+    // convert dict to list
+    var textures_list: [MTLTexture] = []
+    for i in 0..<urls.count {
+        
+        // ensure thread-safety
+        try access_queue.sync {
+            
+            // check whether the images have been loaded successfully
+            if let texture = textures_dict[i] {
+                textures_list.append(texture)
+            } else {
+                throw ImageIOError.load_error
+            }
+        }
+    }
+    
+    return (textures_list, mosaic_pettern_width!)
+}
+
 
 
 // https://stackoverflow.com/questions/26971240/how-do-i-run-a-terminal-command-in-a-swift-script-e-g-xcodebuild
